Attempt to fix additional C pointer interaction

+++ xpc/xpc_darwin.go
@@ -159,7 +159,7 @@ type XpcEventHandler interface {
 func XpcConnect(service string, eh XpcEventHandler) XPC {
 	cservice := C.CString(service)
 	defer C.free(unsafe.Pointer(cservice))
-	return XPC{conn: C.XpcConnect(cservice, unsafe.Pointer(&eh))}
+	return XPC{conn: C.XpcConnect(cservice, C.uintptr_t(uintptr(unsafe.Pointer(&eh))))}
 }
 
 //export handleXpcEvent
